refactor(sampler): consolidate sampling interface, part 1 (#5312)
This PR begins the effort to consolidate PyG's sampling interface in preparation for moving `sample(...)` behind the `GraphStore` interface. This effort is somewhat large in scope and will be broken into multiple PRs for ease of review. There are many TODOs left in this PR; all will be resolved in forthcoming PRs (the code will become cleaner in short order).

This first change moves `NeighborSampler` behind a common base class with clearly defined inputs and outputs. Many more changes will be required before this base class defines a stable interface (e.g. `LinkNeighborSampler` has different semantics, `NeighborSampler.perm_dict` is used in `NeighborLoader`, etc.). But for now, it will act as a first step towards this goal. 

It further refactors `loader/utils.py` into those that are required for sampling (moved into `sampler/utils.py`) and those that are required for filtering (remain in `loader/utils.py`). 

**Note about integration with GraphStore.** Moving sampling behind the graph store requires defining a clear interface for sampling across PyG. Since a graph store should not inherently be coupled with a sampling method used (e.g. one can imagine using a NeighborLoader and LinkNeighborLoader on the same graph store), we must define this interface independently from the GraphStore. One can then imagine the GraphStore defining supported samplers for different loaders (along with the defaults, which sample in-memory), and those samplers are then seamlessly used in the loaders.
