+++ torch_geometric/loader/hgt_loader.py
@@ -5,7 +5,8 @@ from torch import Tensor
 
 from torch_geometric.data import HeteroData
 from torch_geometric.loader.base import DataLoaderIterator
-from torch_geometric.loader.utils import filter_hetero_data, to_hetero_csc
+from torch_geometric.loader.utils import filter_hetero_data
+from torch_geometric.sampler.utils import to_hetero_csc
 from torch_geometric.typing import NodeType
 
 
@@ -125,10 +126,15 @@ class HGTLoader(torch.utils.data.DataLoader):
         self.sample_fn = torch.ops.torch_sparse.hgt_sample
 
         # Convert the graph data into a suitable format for sampling.
+        colptr_dict, row_dict, perm_dict = to_hetero_csc(
+            data, device='cpu', share_memory=kwargs.get('num_workers', 0) > 0)
+
         # NOTE: Since C++ cannot take dictionaries with tuples as key as
         # input, edge type triplets are converted into single strings.
-        self.colptr_dict, self.row_dict, self.perm_dict = to_hetero_csc(
-            data, device='cpu', share_memory=kwargs.get('num_workers', 0) > 0)
+        self.to_rel_type = {key: '__'.join(key) for key in data.edge_types}
+        self.row_dict = remap_keys(row_dict, self.to_rel_type)
+        self.colptr_dict = remap_keys(colptr_dict, self.to_rel_type)
+        self.perm_dict = remap_keys(perm_dict, self.to_rel_type)
 
         super().__init__(input_nodes[1].tolist(), collate_fn=self.collate_fn,
                          **kwargs)
@@ -168,3 +174,10 @@ class HGTLoader(torch.utils.data.DataLoader):
 
     def __repr__(self) -> str:
         return f'{self.__class__.__name__}()'
+
+
+###############################################################################
+
+
+def remap_keys(original: Dict, mapping: Dict) -> Dict:
+    return {mapping[k]: v for k, v in original.items()}
