+++ torch_geometric/loader/utils.py
@@ -1,6 +1,6 @@
 import copy
 import math
-from typing import Dict, Optional, Tuple, Union
+from typing import Dict, Optional, Union
 
 import torch
 from torch import Tensor
@@ -30,72 +30,6 @@ def edge_type_to_str(edge_type: Union[EdgeType, str]) -> str:
     # Since C++ cannot take dictionaries with tuples as key as input, edge type
     # triplets need to be converted into single strings.
     return edge_type if isinstance(edge_type, str) else '__'.join(edge_type)
-
-
-# TODO deprecate when FeatureStore / GraphStore unification is complete
-def to_csc(
-    data: Union[Data, EdgeStorage],
-    device: Optional[torch.device] = None,
-    share_memory: bool = False,
-    is_sorted: bool = False,
-) -> Tuple[Tensor, Tensor, OptTensor]:
-    # Convert the graph data into a suitable format for sampling (CSC format).
-    # Returns the `colptr` and `row` indices of the graph, as well as an
-    # `perm` vector that denotes the permutation of edges.
-    # Since no permutation of edges is applied when using `SparseTensor`,
-    # `perm` can be of type `None`.
-    perm: Optional[Tensor] = None
-
-    if hasattr(data, 'adj'):
-        colptr, row, _ = data.adj.csc()
-
-    elif hasattr(data, 'adj_t'):
-        colptr, row, _ = data.adj_t.csr()
-
-    elif data.edge_index is not None:
-        (row, col) = data.edge_index
-        if not is_sorted:
-            perm = (col * data.size(0)).add_(row).argsort()
-            row = row[perm]
-        colptr = torch.ops.torch_sparse.ind2ptr(col[perm], data.size(1))
-    else:
-        row = torch.empty(0, dtype=torch.long, device=device)
-        colptr = torch.zeros(data.num_nodes + 1, dtype=torch.long,
-                             device=device)
-
-    colptr = colptr.to(device)
-    row = row.to(device)
-    perm = perm.to(device) if perm is not None else None
-
-    if not colptr.is_cuda and share_memory:
-        colptr.share_memory_()
-        row.share_memory_()
-        if perm is not None:
-            perm.share_memory_()
-
-    return colptr, row, perm
-
-
-def to_hetero_csc(
-    data: HeteroData,
-    device: Optional[torch.device] = None,
-    share_memory: bool = False,
-    is_sorted: bool = False,
-) -> Tuple[Dict[str, Tensor], Dict[str, Tensor], Dict[str, OptTensor]]:
-    # Convert the heterogeneous graph data into a suitable format for sampling
-    # (CSC format).
-    # Returns dictionaries holding `colptr` and `row` indices as well as edge
-    # permutations for each edge type, respectively.
-    # Since C++ cannot take dictionaries with tuples as key as input, edge type
-    # triplets are converted into single strings.
-    colptr_dict, row_dict, perm_dict = {}, {}, {}
-
-    for store in data.edge_stores:
-        key = edge_type_to_str(store._key)
-        out = to_csc(store, device, share_memory, is_sorted)
-        colptr_dict[key], row_dict[key], perm_dict[key] = out
-
-    return colptr_dict, row_dict, perm_dict
 
 
 def filter_node_store_(store: NodeStorage, out_store: NodeStorage,
@@ -172,15 +106,25 @@ def filter_hetero_data(
     # edges in `edge` for each node and edge type, respectively:
     out = copy.copy(data)
 
+    # TODO(manan): consolidate behind 'EdgeType':
+    # TODO(manan): remove special handling of 'perm_dict':
+    row_dict_keys = list(row_dict.keys())
+    is_str = len(row_dict_keys) > 0 and isinstance(row_dict_keys[0], str)
+
     for node_type in data.node_types:
         filter_node_store_(data[node_type], out[node_type],
                            node_dict[node_type])
 
     for edge_type in data.edge_types:
-        edge_type_str = edge_type_to_str(edge_type)
-        filter_edge_store_(data[edge_type], out[edge_type],
-                           row_dict[edge_type_str], col_dict[edge_type_str],
-                           edge_dict[edge_type_str], perm_dict[edge_type_str])
+        edge_type_str = edge_type_to_str(edge_type) if is_str else edge_type
+        filter_edge_store_(
+            data[edge_type],
+            out[edge_type],
+            row_dict[edge_type_str],
+            col_dict[edge_type_str],
+            edge_dict[edge_type_str],
+            perm_dict[edge_type_to_str(edge_type)],
+        )
 
     return out
 
@@ -200,10 +144,14 @@ def filter_custom_store(
     # Construct a new `HeteroData` object:
     data = HeteroData()
 
+    # TODO(manan): consolidate behind 'EdgeType':
+    row_dict_keys = list(row_dict.keys())
+    is_str = len(row_dict_keys) > 0 and isinstance(row_dict_keys[0], str)
+
     # Filter edge storage:
     # TODO support edge attributes
     for attr in graph_store.get_all_edge_attrs():
-        key = edge_type_to_str(attr.edge_type)
+        key = edge_type_to_str(attr.edge_type) if is_str else attr.edge_type
         if key in row_dict and key in col_dict:
             edge_index = torch.stack([row_dict[key], col_dict[key]], dim=0)
             data[attr.edge_type].edge_index = edge_index
