+++ torch_geometric/sampler/__init__.py
@@ -1,0 +1,7 @@
+from .base import BaseSampler
+from .neighbor_sampler import NeighborSampler
+
+__all__ = [
+    'BaseSampler',
+    'NeighborSampler',
+]
